+++ sentry-jul/src/main/java/io/sentry/jul/SentryHandler.java
@@ -237,13 +237,10 @@ public final class SentryHandler extends Handler {
   private @NotNull SdkVersion createSdkVersion(final @NotNull SentryOptions sentryOptions) {
     SdkVersion sdkVersion = sentryOptions.getSdkVersion();
 
-    if (sdkVersion == null) {
-      sdkVersion = new SdkVersion();
-    }
-
-    sdkVersion.setName(BuildConfig.SENTRY_JUL_SDK_NAME);
+    final String name = BuildConfig.SENTRY_JUL_SDK_NAME;
     final String version = BuildConfig.VERSION_NAME;
-    sdkVersion.setVersion(version);
+
+    sdkVersion = SdkVersion.updateSdkVersion(sdkVersion, name, version);
     sdkVersion.addPackage("maven:sentry-jul", version);
 
     return sdkVersion;