+++ sentry-logback/src/main/java/io/sentry/logback/SentryAppender.java
@@ -140,13 +140,10 @@ public final class SentryAppender extends UnsynchronizedAppenderBase<ILoggingEve
   private @NotNull SdkVersion createSdkVersion(@NotNull SentryOptions sentryOptions) {
     SdkVersion sdkVersion = sentryOptions.getSdkVersion();
 
-    if (sdkVersion == null) {
-      sdkVersion = new SdkVersion();
-    }
-
-    sdkVersion.setName(BuildConfig.SENTRY_LOGBACK_SDK_NAME);
+    final String name = BuildConfig.SENTRY_LOGBACK_SDK_NAME;
     final String version = BuildConfig.VERSION_NAME;
-    sdkVersion.setVersion(version);
+    sdkVersion = SdkVersion.updateSdkVersion(sdkVersion, name, version);
+
     sdkVersion.addPackage("maven:sentry-logback", version);
 
     return sdkVersion;