+++ sentry-spring-boot-starter/src/main/java/io/sentry/spring/boot/SentryAutoConfiguration.java
@@ -237,13 +237,10 @@ public class SentryAutoConfiguration {
         final @NotNull SentryOptions sentryOptions) {
       SdkVersion sdkVersion = sentryOptions.getSdkVersion();
 
-      if (sdkVersion == null) {
-        sdkVersion = new SdkVersion();
-      }
-
-      sdkVersion.setName(BuildConfig.SENTRY_SPRING_BOOT_SDK_NAME);
+      final String name = BuildConfig.SENTRY_SPRING_BOOT_SDK_NAME;
       final String version = BuildConfig.VERSION_NAME;
-      sdkVersion.setVersion(version);
+      sdkVersion = SdkVersion.updateSdkVersion(sdkVersion, name, version);
+
       sdkVersion.addPackage("maven:sentry-spring-boot-starter", version);
 
       return sdkVersion;