+++ sentry-spring/src/main/java/io/sentry/spring/SentryHubRegistrar.java
@@ -73,13 +73,10 @@ public class SentryHubRegistrar implements ImportBeanDefinitionRegistrar {
     final SentryOptions defaultOptions = new SentryOptions();
     SdkVersion sdkVersion = defaultOptions.getSdkVersion();
 
-    if (sdkVersion == null) {
-      sdkVersion = new SdkVersion();
-    }
-
-    sdkVersion.setName(BuildConfig.SENTRY_SPRING_SDK_NAME);
+    final String name = BuildConfig.SENTRY_SPRING_SDK_NAME;
     final String version = BuildConfig.VERSION_NAME;
-    sdkVersion.setVersion(version);
+    sdkVersion = SdkVersion.updateSdkVersion(sdkVersion, name, version);
+
     sdkVersion.addPackage("maven:sentry-spring", version);
 
     return sdkVersion;