+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # vNext
 
+* Improve Kotlin compatibility for SdkVersion
 * Feat: Support logging via JUL (#1211)
 
 # 4.0.0