+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # vNext
 
 * Feat: Support logging via JUL (#1211)
+* Fix: Fix returning Sentry trace header from Span (#1217)
 
 # 4.0.0